@@ -25,16 +25,17 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "Typing :: Typed",
 ]
-urls = { Homepage = "https://pet.tno.nl/", Documentation = "https://docs.pet.tno.nl/mpc/mpyc/stubs/2.6.2", Source = "https://github.com/TNO-Mpc/mpyc.stubs" }
+urls = { Homepage = "https://pet.tno.nl/", Documentation = "https://docs.pet.tno.nl/mpc/mpyc/stubs/2.11.0", Source = "https://github.com/TNO-MPC/mpyc.stubs" }
 dynamic = ["version"]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-    "mpyc~=0.9.0",
+    "mpyc~=0.10.0",
     "typing_extensions; python_version<'3.11'",
 ]
 
 [project.optional-dependencies]
 tests = [
+    "numpy>=1.0,<3.0",
     "pytest>=8.1",
     "pytest-asyncio",
 ]
@@ -76,7 +77,7 @@ namespace_packages = true
 explicit_package_bases = true
 
 [tool.tbump.version]
-current = "2.6.2"
+current = "2.11.0"
 regex = '''
 \d+\.\d+\.\d+(-(.*))?
 '''
